+++ src/dyntrace/dyntraced/message.hpp
@@ -46,21 +46,21 @@ namespace dyntrace::comm
 
     protected:
         virtual void on_message(const message_type& msg) = 0;
-        virtual void on_error(uint64_t seq, const std::exception* e) = 0;
+        virtual void on_error(uint32_t seq, const std::exception* e) = 0;
 
-        void on_bad_message(uint64_t seq)
+        void on_bad_message(uint32_t seq)
         {
             bad_message_error err;
             on_error(seq, &err);
         }
 
-        uint64_t next_seq() const noexcept
+        uint32_t next_seq() const noexcept
         {
             return _next_seq.fetch_add(1, std::memory_order_relaxed);
         }
 
     private:
-        mutable std::atomic<uint64_t> _next_seq{1};
+        mutable std::atomic<uint32_t> _next_seq{1};
 
         void do_receive()
         {
